Fix LineNum type

+++ samples/IntegratedEventDriven.Samples/Models/IntegratedModels.cs
@@ -461,7 +461,7 @@ public class TSI_ProdBOMLine
     public string ItemId { get; set; } = string.Empty;
 
     [JsonPropertyName("LineNum")]
-    public int LineNum { get; set; }
+    public decimal LineNum { get; set; }
 
     [JsonPropertyName("BOMQty")]
     public decimal BOMQty { get; set; }
